scripts/pyproject.toml
[project]
name = "product-catalog-scripts"
version = "0.1.0"
description = "Scripts for managing product catalog data"
requires-python = ">=3.11"
dependencies = [
    "psycopg[binary]>=3.2.6",
    "python-dotenv>=1.1.0",
]